refactor: delete consentmanagementfilter package
