--- processor/consentmanagementfilter/consentmanagementfilter_test.go
@@ -1,329 +1,0 @@
-package consentmanagementfilter
-
-import (
-	"testing"
-
-	"github.com/stretchr/testify/require"
-
-	backendconfig "github.com/rudderlabs/rudder-server/backend-config"
-	"github.com/rudderlabs/rudder-server/utils/types"
-)
-
-func TestGetConsentManagementInfo(t *testing.T) {
-	type testCaseT struct {
-		description string
-		input       types.SingularEventT
-		expected    ConsentManagementInfo
-	}
-
-	defConsentManagementInfo := ConsentManagementInfo{}
-	testCases := []testCaseT{
-		{
-			description: "should return empty consent management info when no context is sent",
-			input: types.SingularEventT{
-				"anonymousId": "123",
-				"type":        "track",
-				"event":       "test",
-				"properties": map[string]interface{}{
-					"category": "test",
-				},
-			},
-			expected: defConsentManagementInfo,
-		},
-		{
-			description: "should return empty consent management info when no consent management data is sent",
-			input: types.SingularEventT{
-				"anonymousId": "123",
-				"type":        "track",
-				"event":       "test",
-				"properties": map[string]interface{}{
-					"category": "test",
-				},
-				"context": map[string]interface{}{},
-			},
-			expected: defConsentManagementInfo,
-		},
-		{
-			description: "should return empty consent management info when consent management data is not a valid object",
-			input: types.SingularEventT{
-				"anonymousId": "123",
-				"type":        "track",
-				"event":       "test",
-				"properties": map[string]interface{}{
-					"category": "test",
-				},
-				"context": map[string]interface{}{
-					"consentManagement": "not an object",
-				},
-			},
-			expected: defConsentManagementInfo,
-		},
-		{
-			description: "should return empty consent management info when consent management data is malformed",
-			input: types.SingularEventT{
-				"anonymousId": "123",
-				"type":        "track",
-				"event":       "test",
-				"properties": map[string]interface{}{
-					"category": "test",
-				},
-				"context": map[string]interface{}{
-					"consentManagement": map[string]interface{}{
-						"provider": make(chan int),
-					},
-				},
-			},
-			expected: defConsentManagementInfo,
-		},
-		{
-			description: "should return empty consent management info when consent management data is not as per the expected structure",
-			input: types.SingularEventT{
-				"anonymousId": "123",
-				"type":        "track",
-				"event":       "test",
-				"properties": map[string]interface{}{
-					"category": "test",
-				},
-				"context": map[string]interface{}{
-					"consentManagement": map[string]interface{}{
-						"allowedConsentIds": map[string]interface{}{ // this should have been an array
-							"consent": "consent category 1",
-						},
-					},
-				},
-			},
-			expected: defConsentManagementInfo,
-		},
-		{
-			description: "should return consent management info when consent management data is sent correctly",
-			input: types.SingularEventT{
-				"anonymousId": "123",
-				"type":        "track",
-				"event":       "test",
-				"properties": map[string]interface{}{
-					"category": "test",
-				},
-				"context": map[string]interface{}{
-					"consentManagement": map[string]interface{}{
-						"provider": "custom",
-						"allowedConsentIds": []string{
-							"consent category 1",
-							"consent category 2",
-							"",
-							"",
-						},
-						"deniedConsentIds": []string{
-							"consent category 3",
-							"",
-							"consent category 4",
-							"",
-						},
-						"resolutionStrategy": "and",
-						"extra":              "extra field",
-					},
-				},
-			},
-			expected: ConsentManagementInfo{
-				Provider: "custom",
-				AllowedConsentIds: []string{
-					"consent category 1",
-					"consent category 2",
-				},
-				DeniedConsentIds: []string{
-					"consent category 3",
-					"consent category 4",
-				},
-				ResolutionStrategy: "and",
-			},
-		},
-	}
-
-	for _, testCase := range testCases {
-		t.Run(testCase.description, func(t *testing.T) {
-			actual := GetConsentManagementInfo(testCase.input)
-			require.Equal(t, testCase.expected, actual)
-		})
-	}
-}
-
-func TestGetGenericConsentManagementData(t *testing.T) {
-	type testCaseT struct {
-		description string
-		input       *backendconfig.DestinationT
-		expected    map[string]GenericConsentManagementProviderData
-	}
-
-	defGenericConsentManagementData := make(map[string]GenericConsentManagementProviderData)
-	testCases := []testCaseT{
-		{
-			description: "should return empty generic consent management data when no consent management config is present",
-			input: &backendconfig.DestinationT{
-				Config: map[string]interface{}{},
-			},
-			expected: defGenericConsentManagementData,
-		},
-		{
-			description: "should return empty generic consent management data when consent management config is malformed",
-			input: &backendconfig.DestinationT{
-				Config: map[string]interface{}{
-					"consentManagement": make(chan int),
-				},
-			},
-			expected: defGenericConsentManagementData,
-		},
-		{
-			description: "should return empty generic consent management data when consent management config is not as per the expected structure",
-			input: &backendconfig.DestinationT{
-				Config: map[string]interface{}{
-					"consentManagement": map[string]interface{}{ // this should be an array
-						"provider": "dummy",
-					},
-				},
-			},
-			expected: defGenericConsentManagementData,
-		},
-		{
-			description: "should return empty generic consent management data when consent management config entries count is 0",
-			input: &backendconfig.DestinationT{
-				Config: map[string]interface{}{
-					"consentManagement": []interface{}{},
-				},
-			},
-			expected: defGenericConsentManagementData,
-		},
-		{
-			description: "should return generic consent management data when consent management config entries are properly defined",
-			input: &backendconfig.DestinationT{
-				Config: map[string]interface{}{
-					"consentManagement": []interface{}{
-						map[string]interface{}{
-							"provider": "oneTrust",
-							"consents": []map[string]interface{}{
-								{
-									"consent": "consent category 1",
-								},
-								{
-									"consent": "",
-								},
-								{
-									"consent": "consent category 2",
-								},
-							},
-						},
-						map[string]interface{}{
-							"provider": "ketch",
-							"consents": []map[string]interface{}{
-								{
-									"consent": "purpose 1",
-								},
-								{
-									"consent": "",
-								},
-								{
-									"consent": "purpose 2",
-								},
-							},
-						},
-						map[string]interface{}{
-							"provider": "custom",
-							"consents": []map[string]interface{}{
-								{
-									"consent": "custom consent 1",
-								},
-								{
-									"consent": "",
-								},
-								{
-									"consent": "custom consent 2",
-								},
-								{
-									"consent": "",
-								},
-							},
-							"resolutionStrategy": "or",
-						},
-					},
-				},
-			},
-			expected: map[string]GenericConsentManagementProviderData{
-				"oneTrust": {
-					Consents: []string{
-						"consent category 1",
-						"consent category 2",
-					},
-				},
-				"ketch": {
-					Consents: []string{
-						"purpose 1",
-						"purpose 2",
-					},
-				},
-				"custom": {
-					Consents: []string{
-						"custom consent 1",
-						"custom consent 2",
-					},
-					ResolutionStrategy: "or",
-				},
-			},
-		},
-		{
-			description: "should return generic consent management data excluding the providers with empty consents",
-			input: &backendconfig.DestinationT{
-				Config: map[string]interface{}{
-					"consentManagement": []interface{}{
-						map[string]interface{}{
-							"provider": "oneTrust",
-							"consents": []map[string]interface{}{
-								{
-									"consent": "consent category 1",
-								},
-								{
-									"consent": "",
-								},
-								{
-									"consent": "consent category 2",
-								},
-							},
-						},
-						map[string]interface{}{
-							"provider": "ketch",
-							"consents": []map[string]interface{}{
-								{
-									"consent": "",
-								},
-							},
-						},
-						map[string]interface{}{
-							"provider": "custom",
-							"consents": []map[string]interface{}{
-								{
-									"consent": "",
-								},
-								{
-									"consent": "",
-								},
-							},
-							"resolutionStrategy": "or",
-						},
-					},
-				},
-			},
-			expected: map[string]GenericConsentManagementProviderData{
-				"oneTrust": {
-					Consents: []string{
-						"consent category 1",
-						"consent category 2",
-					},
-				},
-			},
-		},
-	}
-
-	for _, testCase := range testCases {
-		t.Run(testCase.description, func(t *testing.T) {
-			actual := GetGenericConsentManagementData(testCase.input)
-			require.Equal(t, testCase.expected, actual)
-		})
-	}
-}
